Reorg tests, allow to deserialize strigs from Decimal
