+++ serde_arrow/src/test_with_arrow/impls/arrow_binary.rs
@@ -144,53 +144,6 @@ fn example_borrowed_nullable() {
         .deserialize_borrowed(&items);
 }
 
-#[cfg(has_arrow_fixed_binary_support)]
-mod fixed_size_binary {
-    use super::*;
-
-    #[test]
-    fn example_vec_as_fixed_size_binary() {
-        let items = [
-            Item(b"foo".to_vec()),
-            Item(b"bar".to_vec()),
-            Item(b"baz".to_vec()),
-        ];
-
-        Test::new()
-            .with_schema(json!([{"name": "item", "data_type": "FixedSizeBinary(3)"}]))
-            .serialize(&items)
-            .deserialize(&items);
-    }
-
-    #[test]
-    fn example_as_fixed_size_binary() {
-        let items = [
-            Item(ByteBuf::from(b"foo")),
-            Item(ByteBuf::from(b"bar")),
-            Item(ByteBuf::from(b"baz")),
-        ];
-
-        Test::new()
-            .with_schema(json!([{"name": "item", "data_type": "FixedSizeBinary(3)"}]))
-            .serialize(&items)
-            .deserialize(&items);
-    }
-
-    #[test]
-    fn example_borrowed_fixed_size_binary() {
-        let items = [
-            Item(Bytes::new(b"foo")),
-            Item(Bytes::new(b"bar")),
-            Item(Bytes::new(b"baz")),
-        ];
-
-        Test::new()
-            .with_schema(json!([{"name": "item", "data_type": "FixedSizeBinary(3)"}]))
-            .serialize(&items)
-            .deserialize_borrowed(&items);
-    }
-}
-
 /// test that byte can be deserialized from string arrays
 mod deserialize_bytes_from_strings {
     use super::*;
